Improve search functionality
