+++ src/Components/MoviesContent/MoviesContent.jsx
@@ -1,5 +1,5 @@
 import React, { useEffect } from "react";
-import { useDispatch, useSelector } from "react-redux";
+import { useDispatch, useSelector, connect } from "react-redux";
 
 import Wrapper from "../Wrapper";
 import Filter from "../Filter";
@@ -8,35 +8,14 @@ import MovieList from "../MovieList";
 import ErrorBoundary from "../ErrorBoundary";
 import { fetchMovies, setSearch } from "../../Redux/actions";
 
-import {
-    BrowserRouter as Router,
-    Link,
-    useLocation
-} from "react-router-dom";
 import MovieEmptyList from "../MovieEmptyList";
 
-function useQuery() {
-    return new URLSearchParams(useLocation().search);
-}
-
-const MoviesContent = () => {
+const MoviesContent = ({sortOrder,filter, searchQuery}) => {
     const dispatch = useDispatch();
     const movies = useSelector(state => state.movies.fetchedMovies);
-
-    const location = useLocation();
-    const route = location.pathname;
-
-    let searchQuery = "";
-
-    if (route === "/search") {
-        const query = useQuery();
-        searchQuery = query.get("q");
-
-        dispatch(setSearch(searchQuery));
-    }
-
+  
     useEffect(() => {
-        dispatch(fetchMovies(searchQuery));
+        dispatch(fetchMovies(searchQuery, filter, sortOrder));
     }, [searchQuery]);
 
     return (
@@ -55,4 +34,12 @@ const MoviesContent = () => {
     )
 };
 
-export default MoviesContent;
+const mapStateToProps = state => {
+    return {
+        sortOrder: state.searchParams.sortOrder,
+        filter: state.searchParams.filter,
+        searchQuery: state.searchParams.search
+    };
+};
+
+export default connect(mapStateToProps, null)(MoviesContent);
